Rename open_slide module to openslide
