+++ tiledbimg/openslide.py
